move _SuperInitCallVisitor to state.py
