+++ torch_sim/optimizers/state.py
@@ -60,57 +60,3 @@ class FireState(OptimState):
 
 
 # there's no GradientDescentState, it's the same as OptimState
-
-
-class _SuperInitCallVisitor(ast.NodeVisitor):
-    """AST visitor that detects calls to parent __init__ implementations."""
-
-    def __init__(self, base_names: set[str]) -> None:
-        self._base_names = base_names
-        self.found = False
-
-    def visit_Call(self, node: ast.Call) -> None:
-        if isinstance(node.func, ast.Attribute) and node.func.attr == "__init__":
-            # Detect super().__init__(...)
-            if isinstance(node.func.value, ast.Call):
-                func = node.func.value.func
-                if isinstance(func, ast.Name) and func.id == "super":
-                    self.found = True
-            # Detect direct BaseClass.__init__(...)
-            elif (
-                isinstance(node.func.value, ast.Name)
-                and node.func.value.id in self._base_names
-            ):
-                self.found = True
-        self.generic_visit(node)
-
-
-def _ensure_init_calls_parent(cls: type) -> None:
-    """Ensure subclasses define __init__ and call a parent initializer."""
-    if "__init__" not in cls.__dict__:
-        raise TypeError(
-            f"{cls.__name__} must define an __init__ method that calls its parent."
-        )
-
-    init_fn = cls.__dict__["__init__"]
-    if not callable(init_fn):
-        raise TypeError(f"{cls.__name__}.__init__ must be callable.")
-
-    try:
-        source = inspect.getsource(init_fn)
-    except (OSError, TypeError) as exc:
-        msg = (
-            f"Cannot verify that {cls.__name__}.__init__ calls its parent. "
-            "Define __init__ in source code and ensure it invokes super().__init__."
-        )
-        raise TypeError(msg) from exc
-
-    tree = ast.parse(textwrap.dedent(source))
-    base_names = {base.__name__ for base in cls.__mro__[1:] if hasattr(base, "__name__")}
-    visitor = _SuperInitCallVisitor(base_names)
-    visitor.visit(tree)
-
-    if not visitor.found:
-        raise TypeError(
-            f"{cls.__name__}.__init__ must call super().__init__ (or a parent class __init__)."
-        )
